frame.* renamed to kinfu_frame.*
