+++ modules/rgbd/src/kinfu_frame.cpp
@@ -5,7 +5,7 @@
 // This code is also subject to the license terms in the LICENSE_KinectFusion.md file found in this module's directory
 
 #include "precomp.hpp"
-#include "frame.hpp"
+#include "kinfu_frame.hpp"
 
 namespace cv {
 namespace kinfu {
