+++ modules/rgbd/src/kinfu_frame.hpp
